+++ chainerrl/replay_buffers/__init__.py
@@ -1,22 +1,7 @@
 from chainerrl.replay_buffers.replay_buffer import ReplayBuffer  # NOQA
-from chainerrl.agents.a3c import A3C  # NOQA
-from chainerrl.agents.acer import ACER  # NOQA
-from chainerrl.agents.al import AL  # NOQA
-from chainerrl.agents.categorical_double_dqn import CategoricalDoubleDQN  # NOQA
-from chainerrl.agents.categorical_dqn import CategoricalDQN  # NOQA
-from chainerrl.agents.ddpg import DDPG  # NOQA
-from chainerrl.agents.double_dqn import DoubleDQN  # NOQA
-from chainerrl.agents.double_pal import DoublePAL  # NOQA
-from chainerrl.agents.dpp import DPP  # NOQA
-from chainerrl.agents.dqn import DQN  # NOQA
-from chainerrl.agents.iqn import IQN  # NOQA
-from chainerrl.agents.nsq import NSQ  # NOQA
-from chainerrl.agents.pal import PAL  # NOQA
-from chainerrl.agents.pcl import PCL  # NOQA
-from chainerrl.agents.pgt import PGT  # NOQA
-from chainerrl.agents.ppo import PPO  # NOQA
-from chainerrl.agents.reinforce import REINFORCE  # NOQA
-from chainerrl.agents.residual_dqn import ResidualDQN  # NOQA
-from chainerrl.agents.sarsa import SARSA  # NOQA
-from chainerrl.agents.td3 import TD3  # NOQA
-from chainerrl.agents.trpo import TRPO  # NOQA
+from chainerrl.replay_buffers.prioritized import PriorityWeightError  # NOQA
+from chainerrl.replay_buffers.prioritized \
+	import PrioritizedReplayBuffer  # NOQA
+from chainerrl.replay_buffers.episodic import EpisodicReplayBuffer  # NOQA
+from chainerrl.replay_buffers.prioritized_episodic \
+	import PrioritizedEpisodicReplayBuffer  # NOQA
