+++ chainerrl/replay_buffers/replay_buffer.py
@@ -1,4 +1,13 @@
+from __future__ import print_function
+from __future__ import unicode_literals
+from __future__ import division
+from __future__ import absolute_import
+from builtins import *  # NOQA
+from future import standard_library
+standard_library.install_aliases()
+
 from chainerrl import replay_buffer
+
 
 class ReplayBuffer(replay_buffer.AbstractReplayBuffer):
